Add conversion method for service catalog SecretTransform
