+++ client/clientset/versioned/typed/appcatalog/v1alpha1/util/appbinding.go
@@ -1,6 +1,7 @@
 package util
 
 import (
+	"bytes"
 	"fmt"
 
 	"github.com/appscode/kutil"
@@ -10,6 +11,8 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/jsonmergepatch"
 	"k8s.io/apimachinery/pkg/util/wait"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/util/jsonpath"
 	api "kmodules.xyz/custom-resources/apis/appcatalog/v1alpha1"
 	cs "kmodules.xyz/custom-resources/client/clientset/versioned/typed/appcatalog/v1alpha1"
 )
@@ -79,4 +82,57 @@ func TryUpdateApp(c cs.AppcatalogV1alpha1Interface, meta metav1.ObjectMeta, tran
 		err = fmt.Errorf("failed to update AppBinding %s/%s after %d attempts due to %v", meta.Namespace, meta.Name, attempt, err)
 	}
 	return
+}
+
+// ref: https://github.com/kubernetes-incubator/service-catalog/blob/37b874716ad709a175e426f5f5638322a600849f/pkg/controller/controller_binding.go#L588
+func TransformCredentials(kc kubernetes.Interface, transforms []api.SecretTransform, credentials map[string]interface{}) error {
+	for _, t := range transforms {
+		switch {
+		case t.AddKey != nil:
+			var value interface{}
+			if t.AddKey.JSONPathExpression != nil {
+				result, err := evaluateJSONPath(*t.AddKey.JSONPathExpression, credentials)
+				if err != nil {
+					return err
+				}
+				value = result
+			} else if t.AddKey.StringValue != nil {
+				value = *t.AddKey.StringValue
+			} else {
+				value = t.AddKey.Value
+			}
+			credentials[t.AddKey.Key] = value
+		case t.RenameKey != nil:
+			value, ok := credentials[t.RenameKey.From]
+			if ok {
+				credentials[t.RenameKey.To] = value
+				delete(credentials, t.RenameKey.From)
+			}
+		case t.AddKeysFrom != nil:
+			secret, err := kc.CoreV1().
+				Secrets(t.AddKeysFrom.SecretRef.Namespace).
+				Get(t.AddKeysFrom.SecretRef.Name, metav1.GetOptions{})
+			if err != nil {
+				return err // TODO: if the Secret doesn't exist yet, can we perform the transform when it does?
+			}
+			for k, v := range secret.Data {
+				credentials[k] = v
+			}
+		case t.RemoveKey != nil:
+			delete(credentials, t.RemoveKey.Key)
+		}
+	}
+	return nil
+}
+
+func evaluateJSONPath(jsonPath string, credentials map[string]interface{}) (string, error) {
+	j := jsonpath.New("expression")
+	buf := new(bytes.Buffer)
+	if err := j.Parse(jsonPath); err != nil {
+		return "", err
+	}
+	if err := j.Execute(buf, credentials); err != nil {
+		return "", err
+	}
+	return buf.String(), nil
 }
